[19] refactor and rename JsonResponse class
